Remove service worker / local realm implementation
Part of an effort to focus and simplify the Boxel developer experience
