+++ packages/host/.gitignore
@@ -3,7 +3,6 @@
 # compiled output
 /dist/
 /tmp/
-/public/worker.js
 
 # dependencies
 /bower_components/
