+++ packages/host/app/controllers/card-error.ts
@@ -1,11 +1,7 @@
 import Controller from '@ember/controller';
-import ENV from '@cardstack/host/config/environment';
 import { withPreventDefault } from '../helpers/with-prevent-default';
 
-const { isLocalRealm } = ENV;
-
 export default class CardErrorController extends Controller {
-  isLocalRealm = isLocalRealm;
   model: any;
   withPreventDefault = withPreventDefault;
 }
