+++ packages/host/app/controllers/card.ts
@@ -1,17 +1,14 @@
 import Controller from '@ember/controller';
-import ENV from '@cardstack/host/config/environment';
 import { withPreventDefault } from '../helpers/with-prevent-default';
 import { service } from '@ember/service';
 import type RouterService from '@ember/routing/router-service';
 import { action } from '@ember/object';
 
 import { tracked } from '@glimmer/tracking';
-const { isLocalRealm } = ENV;
 import { ComponentLike } from '@glint/template';
 import { Model } from '@cardstack/host/routes/card';
 
 export default class CardController extends Controller {
-  isLocalRealm = isLocalRealm;
   isolatedCardComponent: ComponentLike | undefined;
   withPreventDefault = withPreventDefault;
   @service declare router: RouterService;
