+++ packages/host/app/services/card-service.ts
@@ -1,7 +1,6 @@
 import Service, { service } from '@ember/service';
 import { stringify } from 'qs';
 import type LoaderService from './loader-service';
-import type LocalRealm from '../services/local-realm';
 import {
   SupportedMimeType,
   type LooseCardResource,
@@ -21,11 +20,10 @@ import type {
 import type * as CardAPI from 'https://cardstack.com/base/card-api';
 import ENV from '@cardstack/host/config/environment';
 
-const { ownRealmURL, isLocalRealm } = ENV;
+const { ownRealmURL } = ENV;
 
 export default class CardService extends Service {
   @service declare loaderService: LoaderService;
-  @service declare localRealm: LocalRealm;
 
   private apiModule = importResource(
     this,
@@ -49,7 +47,7 @@ export default class CardService extends Service {
   // Note that this should be the unresolved URL and that we need to rely on our
   // fetch to do any URL resolution.
   get defaultURL(): URL {
-    return isLocalRealm ? this.localRealm.url : new URL(ownRealmURL);
+    return new URL(ownRealmURL);
   }
 
   private async fetchJSON(
