+++ packages/host/app/services/local-indexer.ts
@@ -1,0 +1,29 @@
+import Service from '@ember/service';
+import { type RealmAdapter } from '@cardstack/runtime-common';
+import {
+  SearchEntryWithErrors,
+  type RunState,
+} from '@cardstack/runtime-common/search-index';
+
+// Tests inject an implementation of this service to help perform indexing
+// for the test-realm-adapter
+export default class LocalIndexer extends Service {
+  setup(
+    _fromScratch: (realmURL: URL) => Promise<RunState>,
+    _incremental: (
+      prev: RunState,
+      url: URL,
+      operation: 'update' | 'delete'
+    ) => Promise<RunState>
+  ) {}
+  get adapter(): RealmAdapter {
+    return {} as RealmAdapter;
+  }
+  async setEntry(_url: URL, _entry: SearchEntryWithErrors) {}
+}
+
+declare module '@ember/service' {
+  interface Registry {
+    'local-indexer': LocalIndexer;
+  }
+}
