+++ packages/host/app/templates/card.hbs
@@ -1,29 +1,4 @@
-{{#if this.isLocalRealm}}
-  <InLocalRealm as |realm|>
-    {{#if realm.connected}}
-      <this.getIsolatedComponent />
-
-      <LocalRealmFooter
-        @connected={{realm.connected}}
-        @close={{realm.close}}
-      />
-    {{else if realm.isLoading}}
-      Loading local realm...
-    {{else if realm.isEmpty}}
-      <div class='demo-card'>
-        No local realm connected.
-        <a
-          href="#"
-          {{on 'click' (this.withPreventDefault realm.open)}}
-        >Open a local realm</a>
-      </div>
-    {{else}}
-      <p>Local realm state: {{realm.status}}</p>
-    {{/if}}
-  </InLocalRealm>
-{{else}}
-  <this.getIsolatedComponent />
-{{/if}}
+<this.getIsolatedComponent />
 
 <CodeLink />
