+++ packages/host/tests/helpers/index.gts
@@ -58,7 +58,7 @@ interface Options {
 }
 
 // We use a rendered component to facilitate our indexing (this emulates
-// the work that the service worker renderer is doing), which means that the
+// the work that the Fastboot renderer is doing), which means that the
 // `setupRenderingTest(hooks)` from ember-qunit must be used in your tests.
 export const TestRealm = {
   async create(
@@ -99,10 +99,7 @@ async function makeRenderer() {
   );
 }
 
-// This marries together the 2 sides of the message channel,
-// LocalRealm and MessageHandler, which glosses over the postMessage calls
-class MockLocalRealm extends Service {
-  isAvailable = true;
+class MockLocalIndexer extends Service {
   url = new URL(testRealmURL);
   #adapter: RealmAdapter | undefined;
   #entrySetter: EntrySetter | undefined;
@@ -114,7 +111,7 @@ class MockLocalRealm extends Service {
         operation: 'update' | 'delete'
       ) => Promise<RunState>)
     | undefined;
-  setupIndexing(
+  setup(
     fromScratch: (realmURL: URL) => Promise<RunState>,
     incremental: (
       prev: RunState,
@@ -125,14 +122,14 @@ class MockLocalRealm extends Service {
     this.#fromScratch = fromScratch;
     this.#incremental = incremental;
   }
-  async setupIndexRunner(
+  async configureRunner(
     registerRunner: RunnerRegistration,
     entrySetter: EntrySetter,
     adapter: RealmAdapter
   ) {
     if (!this.#fromScratch || !this.#incremental) {
       throw new Error(
-        `fromScratch/incremental not registered with MockLocalRealm`
+        `fromScratch/incremental not registered with MockLocalIndexer`
       );
     }
     this.#entrySetter = entrySetter;
@@ -144,24 +141,21 @@ class MockLocalRealm extends Service {
   }
   async setEntry(url: URL, entry: SearchEntryWithErrors) {
     if (!this.#entrySetter) {
-      throw new Error(`entrySetter not registered with MockLocalRealm`);
+      throw new Error(`entrySetter not registered with MockLocalIndexer`);
     }
     this.#entrySetter(url, entry);
   }
   get adapter() {
     if (!this.#adapter) {
-      throw new Error(`adapter has not been set on MockLocalRealm`);
+      throw new Error(`adapter has not been set on MockLocalIndexer`);
     }
     return this.#adapter;
   }
-  waitForReadiness() {
-    return Promise.resolve();
-  }
 }
 
-export function setupMockLocalRealm(hooks: NestedHooks) {
+export function setupLocalIndexing(hooks: NestedHooks) {
   hooks.beforeEach(function () {
-    this.owner.register('service:local-realm', MockLocalRealm);
+    this.owner.register('service:local-indexer', MockLocalIndexer);
   });
 }
 
@@ -183,13 +177,13 @@ function makeRealm(
   owner: TestContext['owner'],
   realmURL = testRealmURL
 ) {
-  let localRealm = owner.lookup('service:local-realm') as MockLocalRealm;
+  let localIndexer = owner.lookup('service:local-indexer') as MockLocalIndexer;
   return new Realm(
     realmURL,
     adapter,
     async (optsId) => {
       let { registerRunner, entrySetter } = runnerOptsMgr.getOptions(optsId);
-      await localRealm.setupIndexRunner(registerRunner, entrySetter, adapter);
+      await localIndexer.configureRunner(registerRunner, entrySetter, adapter);
     },
     runnerOptsMgr,
     async () =>
