+++ packages/host/tests/integration/search-index-test.ts
@@ -5,7 +5,7 @@ import {
   testRealmURL,
   cleanWhiteSpace,
   setupCardLogs,
-  setupMockLocalRealm,
+  setupLocalIndexing,
   type CardDocFiles,
 } from '../helpers';
 import { RealmPaths } from '@cardstack/runtime-common/paths';
@@ -25,7 +25,7 @@ const testModuleRealm = 'http://localhost:4202/test/';
 
 module('Integration | search-index', function (hooks) {
   setupRenderingTest(hooks);
-  setupMockLocalRealm(hooks);
+  setupLocalIndexing(hooks);
   setupCardLogs(
     hooks,
     async () => await Loader.import(`${baseRealm.url}card-api`)
