+++ packages/realm-server/package.json
@@ -59,7 +59,7 @@
     "setup:demo-in-deployment:production": "mkdir -p /persistent/demo && cp --verbose --update --recursive ../demo-cards/* /persistent/demo/",
     "setup:base-assets": "ts-node --transpileOnly ./scripts/setup-base.ts",
     "start": "NODE_NO_WARNINGS=1 ts-node --transpileOnly main",
-    "start:base": "pnpm run setup:base-assets && NODE_NO_WARNINGS=1 ts-node --transpileOnly main --port=4201 --path='../base' --fromUrl='https://cardstack.com/base/' --toUrl='/base/' --hostLocalRealm",
+    "start:base": "pnpm run setup:base-assets && NODE_NO_WARNINGS=1 ts-node --transpileOnly main --port=4201 --path='../base' --fromUrl='https://cardstack.com/base/' --toUrl='/base/'",
     "start:test-realms": "NODE_NO_WARNINGS=1 ts-node --transpileOnly main --port=4202 --path='./tests/cards' --fromUrl='/node-test/' --toUrl='/node-test/' --path='../host/tests/cards' --fromUrl='/test/' --toUrl='/test/' --fromUrl='https://cardstack.com/base/' --toUrl='http://localhost:4201/base/' --useTestingDomain",
     "start:base:root": "NODE_NO_WARNINGS=1 ts-node --transpileOnly main --port=4203 --path='../base' --fromUrl='https://cardstack.com/base/' --toUrl='/'",
     "start:demo": "NODE_NO_WARNINGS=1 ts-node --transpileOnly main --port=4204 --path='../demo-cards' --fromUrl='/' --toUrl='/' --fromUrl='https://cardstack.com/base/' --toUrl='http://localhost:4201/base/'",
