--- packages/worker/.gitignore
@@ -1,3 +1,0 @@
-node_modules
-/dist
-*.log
