--- packages/worker/babel.config.json
@@ -1,5 +1,0 @@
-{
-  "plugins": [
-    ["@babel/plugin-transform-typescript", { "allowDeclareFields": true }]
-  ]
-}
