--- packages/worker/src/message-handler.ts
@@ -1,151 +1,0 @@
-import { isClientMessage, send } from './messages';
-import assertNever from 'assert-never';
-import { Deferred } from '@cardstack/runtime-common/deferred';
-import type { FetchHandler } from './fetch';
-import {
-  serializeRunState,
-  deserializeRunState,
-  type EntrySetter,
-  type RunnerRegistration,
-  type RunState,
-} from '@cardstack/runtime-common/search-index';
-
-export class MessageHandler {
-  fs: FileSystemDirectoryHandle | null = null;
-  startingUp: Promise<void>;
-  private finishedStarting!: () => void;
-  private entrySetter: EntrySetter | undefined;
-  private source: Client | ServiceWorker | MessagePort | undefined | null;
-  private fromScratchDeferred: Deferred<RunState> | undefined;
-  private incrementalDeferred: Deferred<RunState> | undefined;
-
-  constructor(
-    worker: ServiceWorkerGlobalScope,
-    private fetchHandler: FetchHandler
-  ) {
-    this.startingUp = new Promise((res) => (this.finishedStarting = res));
-    worker.addEventListener('message', (event) => {
-      this.handle(event);
-    });
-  }
-
-  async setupIndexRunner(
-    registerRunner: RunnerRegistration,
-    entrySetter: EntrySetter
-  ) {
-    this.entrySetter = entrySetter;
-    await registerRunner(
-      this.fromScratch.bind(this),
-      this.incremental.bind(this)
-    );
-  }
-
-  async handle(event: ExtendableMessageEvent) {
-    let { data, source } = event;
-    if (!isClientMessage(data) || !source) {
-      return;
-    }
-    this.source = source;
-    switch (data.type) {
-      case 'requestDirectoryHandle':
-        {
-          this.fetchHandler.setRealmsServed(data.realmsServed);
-          send(source, {
-            type: 'directoryHandleResponse',
-            handle: this.fs,
-          });
-        }
-        return;
-      case 'setDirectoryHandle':
-        {
-          this.fs = data.handle;
-          this.fetchHandler.setRealmsServed(data.realmsServed);
-          this.finishedStarting();
-          if (this.fs) {
-            send(source, {
-              type: 'setDirectoryHandleAcknowledged',
-            });
-          }
-        }
-        return;
-      case 'waitForRealmReadiness':
-        {
-          await this.fetchHandler.waitForReadiness();
-          send(source, {
-            type: 'realmReady',
-          });
-        }
-        return;
-      case 'setEntry':
-        {
-          if (!this.entrySetter) {
-            throw new Error(
-              `no entrySetter provided in MessageHandler.setup()`
-            );
-          }
-          let { url, entry } = data;
-          this.entrySetter(new URL(url), entry);
-          send(source, { type: 'setEntryAcknowledged' });
-        }
-        return;
-      case 'fromScratchCompleted':
-        {
-          if (!this.fromScratchDeferred) {
-            throw new Error(
-              `received from scratch index completion response without corresponding request`
-            );
-          }
-          let { state } = data;
-          this.fromScratchDeferred.fulfill(deserializeRunState(state));
-        }
-        return;
-      case 'incrementalCompleted':
-        {
-          if (!this.incrementalDeferred) {
-            throw new Error(
-              `received incremental index completion response without corresponding request`
-            );
-          }
-          let { state } = data;
-          this.incrementalDeferred.fulfill(deserializeRunState(state));
-        }
-        return;
-      default:
-        throw assertNever(data);
-    }
-  }
-
-  private async fromScratch(realmURL: URL): Promise<RunState> {
-    if (!this.source) {
-      throw new Error(
-        `Can't perform fromScratch indexing, the service worker doesn't know which DOM to talk to`
-      );
-    }
-    this.fromScratchDeferred = new Deferred();
-    send(this.source, {
-      type: 'startFromScratch',
-      realmURL: realmURL.href,
-    });
-    return this.fromScratchDeferred.promise;
-  }
-
-  private async incremental(
-    prev: RunState,
-    url: URL,
-    operation: 'update' | 'delete'
-  ): Promise<RunState> {
-    if (!this.source) {
-      throw new Error(
-        `Can't perform incremental indexing, the service worker doesn't know which DOM to talk to`
-      );
-    }
-    this.incrementalDeferred = new Deferred();
-    send(this.source, {
-      type: 'startIncremental',
-      prev: serializeRunState(prev),
-      url: url.href,
-      operation,
-    });
-    return this.incrementalDeferred.promise;
-  }
-}
