--- packages/worker/src/setup-logger.ts
@@ -1,3 +1,0 @@
-import { makeLogDefinitions } from '@cardstack/runtime-common';
-
-(globalThis as any)._logDefinitions = makeLogDefinitions('*=info'); // not really sure of a good place to configure this...
