--- packages/worker/src/util.ts
@@ -1,15 +1,0 @@
-export function timeout(ms: number): Promise<void> {
-  return new Promise((resolve) => setTimeout(() => resolve(), ms));
-}
-
-export async function readFileAsText(
-  handle: FileSystemFileHandle
-): Promise<string> {
-  let file = await handle.getFile();
-  let reader = new FileReader();
-  return await new Promise<string>((resolve, reject) => {
-    reader.onload = () => resolve(reader.result as string);
-    reader.onerror = reject;
-    reader.readAsText(file);
-  });
-}
