--- packages/worker/tsconfig.json
@@ -1,44 +1,0 @@
-{
-  "compilerOptions": {
-    "declaration": true,
-    "esModuleInterop": true,
-    "resolveJsonModule": true,
-    "module": "es2020",
-    "moduleResolution": "Node",
-    "pretty": true,
-    "target": "ES6",
-    "strict": true,
-    "sourceMap": true,
-    "experimentalDecorators": true,
-    "allowUnreachableCode": false,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "noImplicitAny": true,
-    "noImplicitThis": true,
-    "strictNullChecks": true,
-    "strictPropertyInitialization": true,
-    "noFallthroughCasesInSwitch": true,
-    "noImplicitReturns": true,
-    "skipLibCheck": true,
-    "noEmit": true,
-    "lib": ["ES2020", "ScriptHost", "WebWorker", "WebWorker.ImportScripts"],
-    "paths": {
-      "https://cardstack.com/base/*": ["../base/*"],
-      "@cardstack/boxel-ui": ["../boxel-ui/addon"],
-      "@cardstack/boxel-ui/*": ["../boxel-ui/addon/*"]
-    }
-  },
-  "glint": {
-    "environment": ["ember-loose", "ember-template-imports"],
-    "transform": {
-      "include": [
-        "../base/**",
-        "src/**",
-        "../boxel-ui/addon",
-        "../boxel-ui/addon/**"
-      ]
-    }
-  },
-  "include": ["../base/**/*", "src/**/*", "types/**/*"],
-  "exclude": ["../base/__boxel/**/*"]
-}
