--- packages/worker/types/global.d.ts
@@ -1,3 +1,0 @@
-// runtime-common has its own global type declaration that we need to
-// incorporate
-import '../../runtime-common/global';
